+++ src/world/Units/Creatures/Creature.h
@@ -195,7 +195,7 @@ class SERVER_DECL Creature : public Unit
 
         int32 ModDamageDone[SCHOOL_COUNT];
         float ModDamageDonePct[SCHOOL_COUNT];
-        void CalcResistance(uint16 type);
+        void CalcResistance(uint8_t type);
         void CalcStat(uint8_t type);
 
         bool m_canRegenerateHP;
